Create sync modern modules, fix redirect module

See merge request hybrix/hybrixd/node!462

+++ modules/update/module.js
@@ -8,22 +8,24 @@ let execSync = require('child_process').execSync;
 exports.update = update;
 
 function update (proc, data) {
-  execSync('rm -rf ../update/*');
-  execSync('mkdir -p ../update');
+  if (data && data.hasOwnProperty(data.projectId)) {
+    execSync('rm -rf ../update/*');
+    execSync('mkdir -p ../update');
 
-  console.log(' [.] Retrieving archive package.');
-  execSync('curl -OL "https://gitlab.com/api/v4/projects/' + data.projectId + '/repository/archive.tar.gz"');
+    console.log(' [.] Retrieving archive package.');
+    execSync('curl -OL "https://gitlab.com/api/v4/projects/' + data.projectId + '/repository/archive.tar.gz"');
 
-  console.log(' [.] Unpacking archive.');
-  execSync('tar xvf archive.tar.gz -C ../update');
-  execSync('rm -f archive.tar.gz');
+    console.log(' [.] Unpacking archive.');
+    execSync('tar xvf archive.tar.gz -C ../update');
+    execSync('rm -f archive.tar.gz');
 
-  console.log(' [.] Updating files.');
-  execSync('mv -f ../update/*/* ..');
+    console.log(' [.] Updating files.');
+    execSync('mv -f ../update/*/* ..');
 
-  console.log(' [.] Clean up.');
-  execSync('rm -f archive.tar.gz');
-  execSync('rm -rf ../update/*');
+    console.log(' [.] Clean up.');
+    execSync('rm -f archive.tar.gz');
+    execSync('rm -rf ../update/*');
 
-  proc.pass('Update succesfull');
+    proc.pass('Update succesfull');
+  }
 }
